+++ Sources/ZIPFoundation/Entry+Serialization.swift
@@ -2,7 +2,7 @@
 //  Entry+Serialization.swift
 //  ZIPFoundation
 //
-//  Copyright © 2017-2024 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
+//  Copyright © 2017-2026 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
 //  Released under the MIT License.
 //
 //  See https://github.com/weichsel/ZIPFoundation/blob/master/LICENSE for license information.
@@ -62,10 +62,15 @@ extension Entry.LocalFileHeader {
         subRangeStart += Int(self.fileNameLength)
         subRangeEnd = subRangeStart + Int(self.extraFieldLength)
         self.extraFieldData = additionalData.subdata(in: subRangeStart..<subRangeEnd)
+        var extraFields: [ExtensibleDataField] = []
         if let zip64ExtendedInformation = Entry.ZIP64ExtendedInformation.scanForZIP64Field(in: self.extraFieldData,
                                                                                            fields: self.validFields) {
-            self.extraFields = [zip64ExtendedInformation]
+            extraFields.append(zip64ExtendedInformation)
         }
+        if let infoZIPUnicodePath = Entry.InfoZIPUnicodePath.scanForUnicodePath(in: self.extraFieldData) {
+            extraFields.append(infoZIPUnicodePath)
+        }
+        self.extraFields = extraFields
     }
 }