+++ Tests/ZIPFoundationTests/ZIPFoundationErrorConditionTests+ZIP64.swift
@@ -2,7 +2,7 @@
 //  ZIPFoundationErrorConditionTests+ZIP64.swift
 //  ZIPFoundation
 //
-//  Copyright © 2017-2024 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
+//  Copyright © 2017-2026 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
 //  Released under the MIT License.
 //
 //  See https://github.com/weichsel/ZIPFoundation/blob/master/LICENSE for license information.
@@ -78,12 +78,10 @@ extension ZIPFoundationTests {
                                  0xb0, 0x11, 0x00, 0x00, 0x00, 0x00]
         guard let cds = await Entry.CentralDirectoryStructure(data: Data(cdsBytes),
                                                         additionalDataProvider: { count -> Data in
-                                                            guard let pathData = "/".data(using: .utf8) else {
-                                                                throw AdditionalDataError.encodingError
-                                                            }
-                                                            XCTAssert(count == pathData.count)
-                                                            return pathData
-                                                        }) else {
+            let pathData = Data("/".utf8)
+            XCTAssert(count == pathData.count)
+            return pathData
+        }) else {
             XCTFail("Failed to read central directory structure.")
             return nil
         }