+++ Tests/ZIPFoundationTests/ZIPFoundationTests.swift
@@ -2,7 +2,7 @@
 //  ZIPFoundationTests.swift
 //  ZIPFoundation
 //
-//  Copyright © 2017-2024 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
+//  Copyright © 2017-2026 Thomas Zoechling, https://www.peakstep.com and the ZIP Foundation project authors.
 //  Released under the MIT License.
 //
 //  See https://github.com/weichsel/ZIPFoundation/blob/master/LICENSE for license information.
@@ -297,7 +297,11 @@ extension ZIPFoundationTests {
             ("testWriteLargeChunk", testWriteLargeChunk),
             ("testExtractUncompressedZIP64Entries", testExtractUncompressedZIP64Entries),
             ("testExtractCompressedZIP64Entries", testExtractCompressedZIP64Entries),
-            ("testExtractEntryWithZIP64DataDescriptor", testExtractEntryWithZIP64DataDescriptor)
+            ("testExtractEntryWithZIP64DataDescriptor", testExtractEntryWithZIP64DataDescriptor),
+            ("testUnzipSymlink", testUnzipSymlink),
+            ("testUnzipCompressedSymlink", testUnzipCompressedSymlink),
+            ("testEntryScanForInfoZIP", testEntryScanForInfoZIP),
+            ("testInfoZIPUnicodePath", testInfoZIPUnicodePath)
         ]
     }
 
@@ -323,8 +327,16 @@ extension ZIPFoundationTests {
             // Fails for Swift < 4.2 on Linux. We can re-enable that when we drop Swift 4.x support
             ("testZipItemErrorConditions", testZipItemErrorConditions),
             // Applying permissions on symlinks is only relevant on Darwin platforms
-            ("testSymlinkPermissionsTransferErrorConditions", testSymlinkPermissionsTransferErrorConditions),
-            // Applying file modification dates is currently unsupported in corelibs Foundation
+            ("testSymlinkPermissionsTransferErrorConditions", testSymlinkPermissionsTransferErrorConditions)
+        ]
+        #else
+        return []
+        #endif
+    }
+
+    static var posixOnlyTests: [(String, (ZIPFoundationTests) -> () throws -> Void)] {
+        #if os(macOS) || os(iOS) || os(tvOS) || os(visionOS) || os(watchOS) || os(Linux)
+        return [
             ("testSymlinkModificationDateTransferErrorConditions", testSymlinkModificationDateTransferErrorConditions)
         ]
         #else
@@ -366,6 +378,7 @@ extension Data {
 
 #if os(macOS)
 extension NSUserScriptTask {
+
     static func makeVolumeCreationTask(at tempDir: URL, volumeName: String) throws -> NSUserScriptTask {
         let scriptURL = tempDir.appendingPathComponent("createVol.sh", isDirectory: false)
         let dmgURL = tempDir.appendingPathComponent(volumeName).appendingPathExtension("dmg")